T361: DRY — extract shared isPidRunning helper
- _is-pid-running.js: shared helper used by session-cleanup.js and
  session-collision-detector.js (was duplicated 15 lines in each)
- Underscore prefix convention: _helper.js files are utilities, not
  modules. load-modules.js and test-modules.sh now skip _ files.
- All tests pass (8/8 collision, existing suites unaffected)
