+++ run-modules/SessionStart/session-cleanup.js
@@ -6,7 +6,7 @@
 var fs = require("fs");
 var path = require("path");
 var os = require("os");
-var cp = require("child_process");
+var isPidRunning = require("./_is-pid-running");
 
 var TMP = os.tmpdir();
 var PREFIX = ".claude-";
@@ -18,24 +18,6 @@ var STATE_NAMES = [
   "self-analyze-cooldown-",
   "bash-failures-"
 ];
-
-function isPidRunning(pid) {
-  try {
-    if (process.platform === "win32") {
-      var out = cp.execSync("tasklist /FI \"PID eq " + pid + "\" /NH", {
-        encoding: "utf-8",
-        timeout: 3000,
-        stdio: ["pipe", "pipe", "pipe"]
-      });
-      return out.indexOf("" + pid) !== -1;
-    } else {
-      process.kill(pid, 0);
-      return true;
-    }
-  } catch (e) {
-    return false;
-  }
-}
 
 module.exports = function() {
   var cleaned = 0;
